readme: Add Technologies e Tools I use section
This information makes any technical interview unnecessary so you
can be hired instantly by any company.

Since your skills are listed in a README.md file, and it's on the
internet, well you know the math, it's true and proven that you
know very well (up to 5 stars of rating) all of the technologies
listed here.